moved functionality of team generating from TeamsModel class to TeamsGenerator view model still needs to be tested
