+++ Project/Views/TeamGenerator/GenerateTeams.cshtml
@@ -1,4 +1,4 @@
-﻿@model Portfolio.Models.TeamsModel
+﻿@model Portfolio.ViewModels.TeamsGeneratorVM
 
 @{
     ViewData["Title"] = "Generated Teams";
